cmd/reduce: move pkg/testutils/reduce to pkg/cmd/reduce
Release note: None
