+++ pkg/BUILD.bazel
@@ -82,6 +82,8 @@ ALL_TESTS = [
     "//pkg/cmd/prereqs:prereqs_test",
     "//pkg/cmd/publish-artifacts:publish-artifacts_test",
     "//pkg/cmd/publish-provisional-artifacts:publish-provisional-artifacts_test",
+    "//pkg/cmd/reduce/reduce/reducesql:reducesql_test",
+    "//pkg/cmd/reduce/reduce:reduce_test",
     "//pkg/cmd/roachtest/prometheus:prometheus_test",
     "//pkg/cmd/roachtest/tests:tests_test",
     "//pkg/cmd/roachtest:roachtest_test",
@@ -341,8 +343,6 @@ ALL_TESTS = [
     "//pkg/testutils/lint/passes/returnerrcheck:returnerrcheck_test",
     "//pkg/testutils/lint/passes/timer:timer_test",
     "//pkg/testutils/lint/passes/unconvert:unconvert_test",
-    "//pkg/testutils/reduce/reducesql:reducesql_test",
-    "//pkg/testutils/reduce:reduce_test",
     "//pkg/testutils/sqlutils:sqlutils_test",
     "//pkg/testutils/testcluster:testcluster_test",
     "//pkg/testutils/zerofields:zerofields_test",
