+++ pkg/cmd/reduce/reduce/datadriven.go
