+++ pkg/cmd/reduce/reduce/pass.go
