+++ pkg/cmd/reduce/reduce/reducesql/BUILD.bazel
@@ -3,13 +3,13 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "reducesql",
     srcs = ["reducesql.go"],
-    importpath = "github.com/cockroachdb/cockroach/pkg/testutils/reduce/reducesql",
+    importpath = "github.com/cockroachdb/cockroach/pkg/cmd/reduce/reduce/reducesql",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/cmd/reduce/reduce",
         "//pkg/sql/parser",
         "//pkg/sql/sem/builtins",
         "//pkg/sql/sem/tree",
-        "//pkg/testutils/reduce",
     ],
 )
 
@@ -21,9 +21,9 @@ go_test(
     deps = [
         ":reducesql",
         "//pkg/base",
+        "//pkg/cmd/reduce/reduce",
         "//pkg/security",
         "//pkg/server",
-        "//pkg/testutils/reduce",
         "//pkg/testutils/skip",
         "@com_github_jackc_pgx_v4//:pgx",
     ],
