+++ pkg/cmd/reduce/reduce/reducesql/reducesql.go
@@ -16,11 +16,11 @@ import (
 	"regexp"
 	"strings"
 
+	"github.com/cockroachdb/cockroach/pkg/cmd/reduce/reduce"
 	"github.com/cockroachdb/cockroach/pkg/sql/parser"
 	// Import builtins.
 	_ "github.com/cockroachdb/cockroach/pkg/sql/sem/builtins"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
-	"github.com/cockroachdb/cockroach/pkg/testutils/reduce"
 )
 
 // SQLPasses is a collection of reduce.Pass interfaces that reduce SQL
