+++ pkg/cmd/reduce/reduce/reducesql/testdata/simple
