+++ pkg/cmd/reduce/reduce/reducesql/testdata/ungrouped-column
