+++ pkg/cmd/reduce/reduce/testdata/go
